Update rename detection code
Procmon appears to have changed from using IRP_MJ_SET_INFORMATION as the
operation for renaming files to using SetRenameInformationFile. This
change updates the logic to look for the new information and parse the
details slightly differently. This is important because otherwise temp
files will be one of the largest categories of writes, which is not
helpful.

The code still supports the old procmon .csv format, although this has
not been tested due to a lack of test files.